+++ src/renderer.rs
@@ -1,30 +1,85 @@
 //! Module related to implementation of the [`MjRenderer`].
-use crate::render_base::{GlState, RenderBase, sync_geoms};
 use crate::wrappers::mj_visualization::MjvScene;
 use crate::wrappers::mj_rendering::MjrContext;
+
+#[cfg(target_os = "linux")]
+use crate::renderer::egl::GlStateEgl;
+
+use crate::vis_common::sync_geoms;
 use crate::builder_setters;
 use crate::prelude::*;
 
-
-
 use bitflags::bitflags;
-use glutin::prelude::PossiblyCurrentGlContext;
-use glutin::surface::GlSurface;
 use png::Encoder;
-use winit::event_loop::EventLoop;
 
 use std::io::{self, BufWriter, ErrorKind, Write};
 use std::fmt::Display;
 use std::error::Error;
 use std::marker::PhantomData;
+use std::num::NonZero;
 use std::ops::Deref;
 use std::path::Path;
 use std::fs::File;
+
+#[cfg(feature = "renderer-winit-fallback")]
+mod universal;
+
+#[cfg(feature = "renderer-winit-fallback")]
+use universal::GlStateWinit;
+
+#[cfg(target_os = "linux")]
+mod egl;
+
 
 const RGB_NOT_FOUND_ERR_STR: &str = "RGB rendering is not enabled (renderer.with_rgb_rendering(true))";
 const DEPTH_NOT_FOUND_ERR_STR: &str = "depth rendering is not enabled (renderer.with_depth_rendering(true))";
 const INVALID_INPUT_SIZE: &str = "the input width and height don't match the renderer's configuration";
 const EXTRA_INTERNAL_VISUAL_GEOMS: u32 = 100;
+
+
+/// GlState enum wrapper. By default, headless implementation will be used
+/// when supported. Only on failure an invisible winit window will be used.
+pub(crate) enum GlState {
+    #[cfg(feature = "renderer-winit-fallback")] Winit(GlStateWinit),
+    #[cfg(target_os = "linux")] Egl(egl::GlStateEgl),
+}
+
+impl GlState {
+    /// Creates a new [`GlState`], which by default tries to use
+    /// an offscreen implementation. As a fallback, winit will be used.
+    pub(crate) fn new(width: NonZero<u32>, height: NonZero<u32>) -> Result<Self, RendererError> {
+        #[cfg(target_os = "linux")]
+        #[allow(unused_variables)]
+        let egl_err = match GlStateEgl::new(width, height) {
+            Ok(egl_state) => return Ok(Self::Egl(egl_state)),
+            Err(e) => e,
+        };
+
+        #[cfg(feature = "renderer-winit-fallback")]
+        match GlStateWinit::new(width, height) {
+            Ok(winit_state) => return Ok(Self::Winit(winit_state)),
+            #[cfg(not(target_os = "linux"))]
+            Err(e) => {
+                return Err(e);
+            },
+
+            #[cfg(target_os = "linux")]
+            _ => {}
+        }
+
+        #[cfg(target_os = "linux")]
+        Err(RendererError::GlutinError(egl_err))
+    }
+
+    pub(crate) fn make_current(&self) -> glutin::error::Result<()> {
+        match self {
+            #[cfg(target_os = "linux")]
+            Self::Egl(egl_state) => egl_state.make_current(),
+            #[cfg(feature = "renderer-winit-fallback")]
+            Self::Winit(winit_state) => winit_state.make_current()
+        }
+    }
+}
 
 
 /// A builder for [`MjRenderer`].
@@ -55,7 +110,8 @@ impl<M: Deref<Target = MjModel> + Clone> MjRendererBuilder<M> {
             width: 0, height: 0,
             num_visual_internal_geom: EXTRA_INTERNAL_VISUAL_GEOMS, num_visual_user_geom: 0,
             rgb: true, depth: false, font_scale: MjtFontScale::mjFONTSCALE_100,
-            camera: MjvCamera::default(), opts: MjvOption::default(), model_type: PhantomData
+            camera: MjvCamera::default(), opts: MjvOption::default(),
+            model_type: PhantomData
         }
     }
 
@@ -103,7 +159,6 @@ which can be configured at the top of the model's XML like so:
         font_scale: MjtFontScale;       "font scale of drawn text (with [MjrContext]).";
         camera: MjvCamera;              "camera used for drawing.";
         opts: MjvOption;                "visualization options.";
-
     }
 
     /// Builds a [`MjRenderer`].
@@ -117,24 +172,7 @@ which can be configured at the top of the model's XML like so:
             width = global.offwidth as u32;
         }
 
-        let mut event_loop = EventLoop::new().map_err(RendererError::EventLoopError)?;
-        let adapter = RenderBase::new(
-            width, height,
-            "".to_string(),
-            &mut event_loop,
-            false  // don't process events
-        );
-
-        /* Initialize the OpenGL related things */
-        if let Some (GlState { window, gl_context, gl_surface }) = 
-            adapter.state.as_ref()
-        {
-            window.set_visible(false);
-            gl_surface.set_swap_interval(gl_context, glutin::surface::SwapInterval::DontWait)
-                .map_err(RendererError::GlutinError)?;
-        }
-
-        event_loop.set_control_flow(winit::event_loop::ControlFlow::Poll);
+        let gl_state = GlState::new(width.try_into().unwrap(), height.try_into().unwrap())?;
 
         // Initialize the rendering context to render to the offscreen buffer.
         let mut context = MjrContext::new(&model);
@@ -156,8 +194,7 @@ which can be configured at the top of the model's XML like so:
         let renderer = MjRenderer {
             scene, user_scene, context, model, camera: self.camera, option: self.opts,
             flags: RendererFlags::empty(), rgb: None, depth: None,
-            width: width as usize, height: height as usize,
-            adapter, _event_loop: event_loop
+            width: width as usize, height: height as usize, gl_state
         }   // These require special care
             .with_rgb_rendering(self.rgb)
             .with_depth_rendering(self.depth);
@@ -182,8 +219,7 @@ pub struct MjRenderer<M: Deref<Target = MjModel> + Clone> {
     model: M,
 
     /* OpenGL */
-    adapter: RenderBase,
-    _event_loop: EventLoop<()>,
+    gl_state: GlState,
 
     /* Configuration */
     camera: MjvCamera,
@@ -499,10 +535,7 @@ impl<M: Deref<Target = MjModel> + Clone> MjRenderer<M> {
     /// Draws the scene to internal arrays.
     /// Use [`MjRenderer::rgb`] or [`MjRenderer::depth`] to obtain the rendered image.
     fn render(&mut self) {
-        let GlState {gl_context, gl_surface, .. }
-            = self.adapter.state.as_ref().unwrap();
-
-        gl_context.make_current(gl_surface).expect("failed to make OpenGL context current");
+        self.gl_state.make_current().expect("failed to make OpenGL context current");
         let vp = MjrRectangle::new(0, 0, self.width as i32, self.height as i32);
         self.scene.render(&vp, &self.context);
 
@@ -533,6 +566,7 @@ impl<M: Deref<Target = MjModel> + Clone> MjRenderer<M> {
 
 #[derive(Debug)]
 pub enum RendererError {
+    #[cfg(feature = "renderer-winit-fallback")]
     EventLoopError(winit::error::EventLoopError),
     GlutinError(glutin::error::Error)
 }
@@ -540,6 +574,7 @@ pub enum RendererError {
 impl Display for RendererError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
+            #[cfg(feature = "renderer-winit-fallback")]
             Self::EventLoopError(e) => write!(f, "event loop failed to initialize: {}", e),
             Self::GlutinError(e) => write!(f, "glutin failed to initialize: {}", e)
         }
@@ -549,6 +584,7 @@ impl Display for RendererError {
 impl Error for RendererError {
     fn source(&self) -> Option<&(dyn Error + 'static)> {
         match self {
+            #[cfg(feature = "renderer-winit-fallback")]
             Self::EventLoopError(e) => Some(e),
             Self::GlutinError(e) => Some(e)
         }